[merge] franchise에 있던 모델 main에 넣음!
