+++ hangyo/config/settings/base.py
@@ -32,7 +32,6 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'main',
-    'franchise',
     'import_export',
 ]
