--- hangyo/franchise/models.py
@@ -1,8 +1,0 @@
-from django.db import models
-
-# Create your models here.
-class stores(models.Model):
-    store_name = models.CharField(max_length=100)
-    phone_num = models.CharField(max_length=100)
-    location = models.CharField(max_length=100)
-    store_type = models.CharField(max_length=100)
