+++ hangyo/main/admin.py
@@ -1,3 +1,11 @@
 from django.contrib import admin
+from .models import Stores
+from import_export.admin import ExportActionModelAdmin, ImportExportMixin, ImportMixin
+
+class StoreAdmin(ImportExportMixin, admin.ModelAdmin):
+    pass
+
+
+admin.site.register(Stores,StoreAdmin)
 
 # Register your models here.
