+++ hangyo/main/models.py
@@ -2,5 +2,8 @@ from django.db import models
 
 # Create your models here.
 
-class Temp(models.Model):
-    temp = models.TextField()
+class Stores(models.Model):
+    store_name = models.CharField(max_length=100)
+    phone_num = models.CharField(max_length=100)
+    location = models.CharField(max_length=100)
+    store_type = models.CharField(max_length=100)
